FILE-411: Moved redirection integration tests into separate class
